Add files via upload
The issue of keeping one's employees happy and satisfied is a perennial and age-old challenge. If an employee you have invested so much time and money leaves for "greener Pastures", then this would mean that you would have to spend even more time and money to hire somebody else.  l